Fixed filter bug and improved reliability of Solver.
A bug existed wherein the endsWith and containsExact fields were swapped. The Solver class has been converted to an enum to ensure that only a singleton ever exists of it (since it loads a lot into memory). The UI now reflects when the initialization of context is complete, so that searches appear fast.